Simplify the unified `HashJoin` change without changing behavior
Fold complexity accreted while the branch was built: `NUM_BUCKETS` goes back
to a constant, which takes the aggregation files out of the change;
`RightTableData` gets one resumable per-worker cursor instead of eleven
open-coded walks; `createKeyGetter` becomes one free template instead of
three copies.

+++ src/Common/HashTable/HashTable.h
@@ -425,7 +425,7 @@ protected:
     template <typename, typename, typename, typename, typename, typename, Int32, typename>
     friend class TwoLevelHashTable;
 
-    template <typename, typename, Int32>
+    template <typename, typename, size_t>
     friend class TwoLevelStringHashTable;
 
     template <typename SubMaps>
